Dev: feed CovidTable with CSV
